# lab1/swagger.yaml
openapi: 3.0.3
info:
  title: Organization Management API
  description: API для управления организацями
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: localhost
  - url: https://se.ifmo.ru/~s368796/
    description: Helios
...
components:
  schemas:
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
          default: 0
          description: Номер страницы
        size:
          type: integer
          format: int32
          minimum: 0
          default: 20
          description: Размер страницы
...
    Coordinates:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: integer
          format: int32
          minimum: -365
          description: Значение должно быть больше -366
        y:
          type: number
          format: double
    Location:
      type: object
      required:
        - x
        - y
        - z
      properties:
        x:
          type: number
          format: float
        y:
          type: number
          format: double
        z:
          type: integer
          format: int64
          description: Поле не может быть null
        name:
          type: string
          nullable: true
    Address:
      type: object
      required:
        - street
        - town
      properties:
        street:
          type: string
          minLength: 1
          description: Строка не может быть пустой
        town:
          $ref: '#/components/schemas/Location'
    Organization:
      type: object
      required:
        - name
        - coordinates
        - annualTurnover
        - type
      properties:
        id:
          type: integer
          readOnly: true
          minimum: 1
          description: Генерируется автоматически, должно быть больше 0
        name:
          type: string
          minLength: 1
          description: Поле не может быть null или пустым
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        creationDate:
          type: string
          format: date
          readOnly: true
          description: Генерируется автоматически
        annualTurnover:
          type: integer
          format: int64
          minimum: 1
          description: Значение должно быть больше 0
        fullName:
          type: string
          maxLength: 918
          nullable: true
        employeesCount:
          type: integer
          minimum: 1
          nullable: true
          description: Значение должно быть больше 0 если указано
        type:
          type: string
          enum: [ PUBLIC, TRUST, OPEN_JOINT_STOCK_COMPANY ]
        postalAddress:
          $ref: '#/components/schemas/Address'
          nullable: true
    OrganizationArray:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        totalCount:
          type: integer
        page:
          type: integer
        size:
          type: integer
...
paths:
...
  /orgdirectory/filter/turnover/{min-annual-turnover}/{max-annual-turnover}:
    post:
      tags:
        - OrgDirectories
      summary: Фильтрация организаций по годовому обороту
      description: Возвращает организации с годовым оборотом в указанном диапазоне
      parameters:
        - name: min-annual-turnover
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: max-annual-turnover
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pagination'
      responses:
        '200':
          description: Успешная фильтрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationArray'
        '400':
          description: Неверный диапазон оборотов
  /orgdirectory/filter/type/{type}:
    post:
      tags:
        - OrgDirectories
      summary: Фильтрация организаций по типу
      description: Возвращает организации указанного типа
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [ PUBLIC, TRUST, OPEN_JOINT_STOCK_COMPANY ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pagination'
      responses:
        '200':
          description: Успешная фильтрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationArray'